Update .list-standout to preserve line height
The arrows for the `.list-standout` style are inserted as inline-block pseudo-elements. Since their `font-size` is set to `1.5em`, they increase the height of the line they're on, disrupting the vertical rhythm of the list.

I suggest removing the arrows from the flow of the document. In my suggestion, I have absolutely positioned them and adjusted the other styles accordingly.